Extract common tests into a shared example
