+++ spec/unit_spec_helper.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'rails'
 
 # load all shared example files
-Dir["./spec/unit/client/**/shared/**/*.rb"].sort.each { |f| require f }
+Dir["./spec/unit/**/shared/**/*.rb"].sort.each { |f| require f }
 
 def unit_example?(metadata)
   metadata[:file_path] =~ %r{spec/unit}
